temporary check-in

+++ app/src/main/java/com/illiad/troad/service/security/Secret.kt
@@ -6,9 +6,9 @@ interface Secret {
 
     val cryptoType: Cryptos?
 
-    val cryptoTypeByte: Byte
+    val cryptoTypeByte: Int
 
-    val cryptoLength: Short
+    val cryptoLength: Int
 
     fun offset(): ByteArray?
 }
